refactor: split pylon/server.ts into 21 route modules (#146)
1,777-line monolith → 220-line composition root + 21 route modules.
Each module exports a factory receiving (RouteDeps, RouteRefs).
All 84 existing tests pass without modification.
